requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1

# API Clients
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.154.0
slack-sdk==3.33.4
requests==2.32.3
anthropic==0.39.0
google-generativeai==0.8.3

# Utilities
python-multipart==0.0.20
aiofiles==24.1.0
httpx==0.28.1

# Logging
structlog==24.4.0
python-json-logger==3.2.1
